+++ src/Event/AppHomeOpenedEvent.php
@@ -1,7 +1,7 @@
 <?php
 declare(strict_types=1);
 
-namespace App\Service\Event;
+namespace App\Event;
 
 use App\Service\UserService;
